payment options update

+++ src/app/modules/paymentOptions/paymentOptions.services.ts
@@ -2,12 +2,14 @@ import { PaymentSystemOptions } from '@prisma/client';
 import httpStatus from 'http-status';
 import ApiError from '../../../errors/ApiError';
 import prisma from '../../../shared/prisma';
+import { IPaymentOptionPayload } from './paymentOptions.constant';
 
 const createNew = async (
   userid: string,
   userRole: string,
-  payload: PaymentSystemOptions,
+  payload: IPaymentOptionPayload,
 ) => {
+  let orgId = null;
   if (userRole === 'STAFF') {
     const userInfo = await prisma.staff.findUnique({
       where: { staffInfoId: userid },
@@ -22,34 +24,38 @@ const createNew = async (
         'Only accounts manager and admin staff and owner can cerate Payment options',
       );
     }
+    orgId = userInfo.organizationId;
+  } else {
+    const userInfo = await prisma.user.findUnique({
+      where: { id: userid },
+    });
+    if (!userInfo) {
+      throw new ApiError(httpStatus.NOT_FOUND, 'User info not found');
+    }
+
+    orgId = userInfo.organizationId;
+  }
+  if (!orgId) {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'Organization info not found');
   }
 
-  const organizationInfo = await prisma.organization.findUnique({
-    where: { ownerId: userid },
-    include: {
-      owner: true,
+  const result = await prisma.paymentSystemOptions.create({
+    data: {
+      paymentCategory: payload.paymentCategory,
+      methodName: payload.methodName,
+      accountNumber: payload.accountNumber,
+      description: payload.description,
+      organizationId: orgId,
     },
   });
-  if (!organizationInfo?.owner?.verified) {
-    throw new ApiError(
-      httpStatus.BAD_REQUEST,
-      'Your organization is not verified',
-    );
-  }
-  const result = await prisma.paymentSystemOptions.create({ data: payload });
   return result;
 };
 
-const getSingle = async (paymentOptionId: string) => {
-  const result = await prisma.paymentSystemOptions.findUnique({
-    where: { id: paymentOptionId },
-    include: {
-      organization: { include: { BusinessType: true, owner: true } },
-    },
+const getSingle = async (organizationId: string) => {
+  const result = await prisma.paymentSystemOptions.findMany({
+    where: { organizationId: organizationId },
   });
-  if (!result) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'Referred code not found');
-  }
+  return result;
 };
 
 const updateSingle = async (
@@ -58,6 +64,7 @@ const updateSingle = async (
   paymentOptionId: string,
   payload: Partial<PaymentSystemOptions>,
 ) => {
+  let orgId = null;
   if (userRole === 'STAFF') {
     const userInfo = await prisma.staff.findUnique({
       where: { staffInfoId: userId },
@@ -72,6 +79,28 @@ const updateSingle = async (
         'Only accounts manager and admin staff and owner can change Payment options',
       );
     }
+    orgId = userInfo.organizationId;
+  } else {
+    const userInfo = await prisma.user.findUnique({
+      where: { id: userId },
+    });
+    if (!userInfo) {
+      throw new ApiError(httpStatus.NOT_FOUND, 'User info not found');
+    }
+
+    orgId = userInfo.organizationId;
+  }
+  if (!orgId) {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'Organization info not found');
+  }
+  const isPaymentOptionsExist = await prisma.paymentSystemOptions.findUnique({
+    where: { id: paymentOptionId },
+  });
+  if (!isPaymentOptionsExist) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Payment options not found');
+  }
+  if (orgId !== isPaymentOptionsExist.organizationId) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Organization id not matched');
   }
   const result = await prisma.paymentSystemOptions.update({
     where: { id: paymentOptionId },
@@ -86,6 +115,7 @@ const deleteSingle = async (
   userRole: string,
   paymentOptionId: string,
 ) => {
+  let orgId = null;
   if (userRole === 'STAFF') {
     const userInfo = await prisma.staff.findUnique({
       where: { staffInfoId: userId },
@@ -97,9 +127,31 @@ const deleteSingle = async (
     if (!validStaff.includes(userInfo.role)) {
       throw new ApiError(
         httpStatus.BAD_REQUEST,
-        'Only accounts manager and admin staff and owner can change Payment options',
+        'Only accounts manager and admin staff and owner can delete Payment options',
       );
     }
+    orgId = userInfo.organizationId;
+  } else {
+    const userInfo = await prisma.user.findUnique({
+      where: { id: userId },
+    });
+    if (!userInfo) {
+      throw new ApiError(httpStatus.NOT_FOUND, 'User info not found');
+    }
+
+    orgId = userInfo.organizationId;
+  }
+  if (!orgId) {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'Organization info not found');
+  }
+  const isPaymentOptionsExist = await prisma.paymentSystemOptions.findUnique({
+    where: { id: paymentOptionId },
+  });
+  if (!isPaymentOptionsExist) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Payment options not found');
+  }
+  if (orgId !== isPaymentOptionsExist.organizationId) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Organization id not matched');
   }
   const result = await prisma.paymentSystemOptions.delete({
     where: { id: paymentOptionId },
